+++ src/controller/post.ts
@@ -1,5 +1,5 @@
 import { IncomingMessage, ServerResponse } from 'http';
-import { validate } from 'uuid';
+import { v4 as uuid} from 'uuid';
 import { correctData } from '../util/correctData';
 import { dataUsers, IUser } from '../types/types';
 import { errorRes, successOk, headerRes } from '../util/response';
@@ -8,44 +8,33 @@ export const post = async (
   url: string,
   req: IncomingMessage,
   res: ServerResponse,
-  userDb: dataUsers
+  userDb: dataUsers,
 ): Promise<void> => {
-  if (url?.startsWith('/api/users/')) {
-    const id = url.split('/')[3];
-    if (!id) {
-    errorRes(res, 404, process.env.USER_ERROR_MESSAGE as string);
-    } else if (!validate(id)) {
-    errorRes(res, 400, process.env.USER_ID_ERROR_MESSAGE as string);
-    } else if (!userDb.find((i) => i.id === id)) {
-    errorRes(res, 404, process.env.USER_ERROR_MESSAGE as string);
-    } else {
-      try {
-        const data: IUser = await correctData(req, res);
-        if (
-          ['username', 'age', 'hobbies'].every((key) => 
-          data.hasOwnProperty(key)) &&
-          Array.isArray(data.hobbies) &&
-          data.username.trim().length > 0
-        ) {
-        errorRes(res, 400, process.env.BODY_ERROR_MESSAGE as string);
-        } else {
-          const { username, age, hobbies } = data;
-          const correctUser: IUser = {
-            id,
-            username: username.trim(),
-            age,
-            hobbies,
-          };
-          const index = userDb.findIndex((i) => i.id === id);
-          userDb[index] = correctUser;
-          res.writeHead(200, headerRes);
-          res.end(JSON.stringify(correctUser));
-        }
-      } catch (err) {
-      errorRes(res, 500, process.env.SERVER_ERROR_MESSAGE as string);
+  if (url === '/api/users') {
+    try {
+      let data: IUser = await correctData(req, res);
+      if (
+        ['username', 'age', 'hobbies'].every((key) => data.hasOwnProperty(key)) &&
+        Array.isArray(data.hobbies) &&
+        data.username.trim().length > 0
+      ) {
+        const { username, age, hobbies } = data;
+        const getUser: IUser = {
+          id: uuid(),
+          username: username.trim(),
+          age,
+          hobbies,
+        };
+        res.writeHead(201, headerRes);
+        userDb.push(getUser);
+        res.end(JSON.stringify(getUser));
+      } else {
+        errorRes(res, 400, process.env.USER_ID_ERROR_MESSAGE as string);
       }
+    } catch (error) {
+      errorRes(res, 500, process.env.SERVER_ERROR_MESSAGE as string);
     }
   } else {
-  errorRes(res, 400, process.env.BODY_ERROR_MESSAGE as string);
+    errorRes(res, 400, process.env.BODY_ERROR_MESSAGE as string);
   }
 };
