+++ src/server/clusterServer.ts
@@ -1,14 +1,17 @@
 import { createServer, IncomingMessage, ServerResponse } from 'http';
+// import { setMaxListeners } from 'node:events';
 import cluster from 'cluster';
 import { cpus } from 'os';
 import { switchServer } from './switchServer';
 import { dataUsers } from '../types/types';
 import 'dotenv/config';
 
+// setMaxListeners(36);
 const SERVER_PORT = process.env.SERVER_PORT || 3001;
 const MULTI_PORT = process.env.MULTI_PORT || 4001;
 const args = process.argv.slice(2);
 let dbSet: dataUsers = [];
+const cpuCount = cpus().length;
 
 const server = createServer((
   req: IncomingMessage, 
@@ -25,7 +28,8 @@ const singlServer = (): void => {
 
 const multiServers = (): void => {
   if (cluster.isPrimary) {
-    console.log(`Master start ${process.pid}`)
+    console.log(`The total number of CPUs is ${cpuCount}`);
+    console.log(`Primary pid ${process.pid}`)
     cpus().forEach((CpuInfo, index) => {
       const port = +MULTI_PORT + index;
       cluster.fork({ MULTI_PORT: port });
@@ -36,7 +40,7 @@ const multiServers = (): void => {
   }
   if (cluster.isWorker) {
     server.listen(process.env.MULTI_PORT, () => {
-      console.log(`Server started ${MULTI_PORT}, Worker start ${process.pid}`);
+      console.log(`Server started ${MULTI_PORT}, worker pid=${process.pid}`);
     });
     process.on('message', (message: dataUsers) => {
       dbSet = message;
